Refactor player movement to handle touch/keyboard input
It works overall, I love how easy it is to handle inputs from different devices and get the same effect!
The problem is that the touch devices first touch is way off, and then it starts to work.
Haven't found a remedy as yet, looks like yet another Unity bug